Binds mysqld to localhost. Wait for start change.
The mysqld launched by the rule now only binds to localhost. This stops annoying firewall messages on os x.

Changes the way we wait for mysql to start. Previously we waited for the socket to accept connections.
When the test system is underload this can result in setup failures as mysqld is not actually ready to
process commands. Now the test rule invokes mysqladmin ping until a successful result is returned.